Russian messages from programms are now shown correctly Relative compilers paths in compilers.ini are supported Automatic detection of tests and checker in problem folder added Delphi compiler (dcc32.exe and some other files) is included and added to config Some small changes & bugfixes

+++ Sources/Tester.cs
@@ -13,36 +13,25 @@ namespace olympchecker_gui
 
     static class Tester
     {
-        public struct Parameters
+        public class Parameters
         {
-            public string source, testsDir, inputFile, outputFile;
+            public string solution, inputFile, outputFile;
+            public List<Pair<string, string>> tests;
             public string checker;
             public int timeLimit;
-            public bool internalCheck, exactCheck, standartIO;
-            public Compiler compiler, compilerChecker;
+            public bool internalCheck, standartIO;
             public ProgressNotifier progressNotifier;
         };
         private enum Code { OK, WA, PE, FL, RE, TL };
         private static readonly string[] resultDirs = { "OK", "WA", "PE", "FL", "RE", "TL" };
-        private static readonly string workDir = "work";
-        private static readonly string solutionExec = "olympSolution.exe";
-        private static readonly string checkerExec = "olympChecker.exe";
         private static readonly string tempOutput = "tmpfile.out";
-        private static List<string> tests;
         private static int testsPassed;
         private static int maxTime;
         private static Parameters parameters;
         private static Thread testingThread;
 
-        static Tester()
-        {
-            Directory.CreateDirectory(workDir);
-            Directory.SetCurrentDirectory(workDir);
-        }
-
         public static void TestSolution(Parameters parameters)
         {
-            tests = new List<string>();
             testsPassed = 0;
             maxTime = 0;
 
@@ -54,6 +43,7 @@ namespace olympchecker_gui
             {
                 parameters.outputFile = "out";
             }
+
             Tester.parameters = parameters;
 
             testingThread = new Thread(doTestSolution);
@@ -77,8 +67,6 @@ namespace olympchecker_gui
             try
             {
                 if (!CleanBefore()) { return; }
-                if (!CompileSolution()) { return; }
-                if (!parameters.internalCheck && !CompileChecker()) { return; }
                 if (!FindTests()) { return; }
                 if (!PerformTesting()) { return; }
                 PrintResult();
@@ -89,7 +77,7 @@ namespace olympchecker_gui
 
         private static void KillSolutions()
         {
-            foreach (Process process in Process.GetProcessesByName(Path.GetFileNameWithoutExtension(solutionExec)))
+            foreach (Process process in Process.GetProcessesByName(Path.GetFileNameWithoutExtension(parameters.solution)))
             {
                 process.Kill();
             }
@@ -105,8 +93,12 @@ namespace olympchecker_gui
             {
                 foreach (string file in Directory.GetFiles("."))
                 {
-                    Utils.WaitForFile(file);
-                    File.Delete(file);
+                    if (Path.GetFileName(file) != Path.GetFileName(parameters.solution) &&
+                        Path.GetFileName(file) != Path.GetFileName(parameters.checker))
+                    {
+                        Utils.WaitForFile(file);
+                        File.Delete(file);
+                    }
                 }
 
                 foreach (string dir in Directory.GetDirectories("."))
@@ -148,60 +140,20 @@ namespace olympchecker_gui
             }
         }
 
-        private static bool CompileSolution()
-        {
-            Utils.AddPreparation("Компилирую решение...");
-            string result = Utils.CopyOrCompile(parameters.source, solutionExec);
-            if (String.IsNullOrEmpty(result))
-            {
-                Utils.AddRes("[OK]", Color.Green);
-                return true;
-            }
-            else
-            {
-                Utils.AddRes("[Ошибка]", Color.Red, result);
-                return false;
-            }
-        }
-
-        private static bool CompileChecker()
-        {
-            Utils.AddPreparation("Компилирую проверяющую программу...");
-            string result = Utils.CopyOrCompile(parameters.checker, checkerExec);
-            if (String.IsNullOrEmpty(result))
-            {
-                Utils.AddRes("[OK]", Color.Green);
-                return true;
-            }
-            else
-            {
-                Utils.AddRes("[Ошибка]", Color.Red, result);
-                return false;
-            }
-        }
-
         private static bool FindTests()
         {
-            Utils.AddPreparation("Поиск тестов...");
-
-            tests = (from file in Directory.GetFiles(parameters.testsDir)
-                     where !file.EndsWith(".a") && File.Exists(file + ".a")
-                     orderby file ascending
-                     select file)
-                    .ToList();
-
-            Utils.AddRes("[" + tests.Count + "]", Color.Green);
-
+            Utils.AddPreparation("Найдено тестов:");
+            Utils.AddRes("[" + parameters.tests.Count + "]", Color.Green);
             return true;
         }
 
         private static bool PerformTesting()
         {
-            foreach (string test in tests)
+            foreach (var test in parameters.tests)
             {
                 Thread.Sleep(200);
 
-                Utils.AddTesting(String.Format("Тест {0}: ", Path.GetFileNameWithoutExtension(test)));
+                Utils.AddTesting(String.Format("Тест {0}: ", Path.GetFileNameWithoutExtension(test.First)));
                 try
                 {
                     PrepareTest(test);
@@ -215,14 +167,14 @@ namespace olympchecker_gui
 
                 if (parameters.progressNotifier != null)
                 {
-                    parameters.progressNotifier(tests.IndexOf(test) + 1, tests.Count);
+                    parameters.progressNotifier(parameters.tests.IndexOf(test) + 1, parameters.tests.Count);
                 }
             }
 
             return true;
         }
 
-        private static void PrepareTest(string test)
+        private static void PrepareTest(Pair<string, string> test)
         {
             if (File.Exists(parameters.inputFile))
             {
@@ -234,8 +186,8 @@ namespace olympchecker_gui
                 Utils.WaitForFile("correct.out");
                 File.Delete("correct.out");
             }
-            File.Copy(test, parameters.inputFile);
-            File.Copy(test + ".a", "correct.out");
+            File.Copy(test.First, parameters.inputFile);
+            File.Copy(test.Second, "correct.out");
 
             File.Delete(tempOutput);
             File.Delete(parameters.outputFile);
@@ -254,9 +206,9 @@ namespace olympchecker_gui
             streams.streamWriter.Close();
         }
 
-        public static void PerformTest(string test)
+        public static void PerformTest(Pair<string, string> test)
         {
-            Process process = Utils.StartProcess(solutionExec, String.Empty, true);
+            Process process = Utils.StartProcess(parameters.solution, String.Empty, true);
 
             if (parameters.standartIO)
             {
@@ -296,6 +248,11 @@ namespace olympchecker_gui
                 maxTime = Math.Max(maxTime, time);
             }
 
+            if (parameters.standartIO && !File.Exists(parameters.outputFile))
+            {
+                Utils.WaitForFile(tempOutput);
+                File.Move(tempOutput, parameters.outputFile);
+            }
 
             Code code = Code.OK;
             string additionalInfo = null;
@@ -311,18 +268,13 @@ namespace olympchecker_gui
             }
             else
             {
-                if (parameters.standartIO && !File.Exists(parameters.outputFile))
-                {
-                    Utils.WaitForFile(tempOutput);
-                    File.Move(tempOutput, parameters.outputFile);
-                }
                 code = CheckAnswer(parameters.inputFile, "correct.out", parameters.outputFile, out additionalInfo);
             }
 
             Directory.CreateDirectory(resultDirs[(int)code]);
-            File.Copy(test, Path.Combine(resultDirs[(int)code], Path.GetFileName(test)), true);
-            File.Copy(test + ".a", Path.Combine(resultDirs[(int)code], Path.GetFileName(test) + ".a"), true);
-            File.Copy(parameters.outputFile, Path.Combine(resultDirs[(int)code], Path.GetFileName(test) + ".out"), true);
+            File.Copy(test.First, Path.Combine(resultDirs[(int)code], Path.GetFileName(test.First)), true);
+            File.Copy(test.Second, Path.Combine(resultDirs[(int)code], Path.GetFileName(test.Second)), true);
+            File.Copy(parameters.outputFile, Path.Combine(resultDirs[(int)code], Path.GetFileName(test.First) + ".out"), true);
 
             PrintTestResult(code, time, additionalInfo);
             if (code == Code.OK)
@@ -335,12 +287,12 @@ namespace olympchecker_gui
         {
             if (parameters.internalCheck)
             {
-                Code code = Utils.CompareFiles(answerFile, outFile, parameters.exactCheck, out additionalInfo) ? Code.OK : Code.WA;
+                Code code = Utils.CompareFiles(answerFile, outFile, out additionalInfo) ? Code.OK : Code.WA;
                 return code;
             }
             else
             {
-                Process process = Utils.StartProcess(checkerExec, inputFile + " " + answerFile + " " + outFile);
+                Process process = Utils.StartProcess(parameters.checker, inputFile + " " + outFile + " " + answerFile);
                 additionalInfo = process.StandardOutput.ReadToEnd();
                 process.WaitForExit();
                 Code code = (Code)process.ExitCode;
@@ -385,9 +337,9 @@ namespace olympchecker_gui
             Utils.AddResult("Тестирование завершено:");
             Color color;
             int score = 0;
-            if (tests.Count != 0)
+            if (parameters.tests.Count != 0)
             {
-                score = 100 * testsPassed / tests.Count;
+                score = 100 * testsPassed / parameters.tests.Count;
             }
             if (score == 100)
             {
@@ -401,7 +353,7 @@ namespace olympchecker_gui
             {
                 color = Color.DarkOrange;
             }
-            Utils.AddRes(testsPassed + "/" + tests.Count + " (" + score + "/100)", color);
+            Utils.AddRes(testsPassed + "/" + parameters.tests.Count + " (" + score + "/100)", color);
 
             Utils.AddResult("Максимальное время (без учёта TL):");
             Utils.AddRes(maxTime + " ms",
